+++ include/siddiqsoft/private/rest_request.hpp
@@ -55,7 +55,6 @@
 #include "siddiqsoft/SplitUri.hpp"
 #include "siddiqsoft/date-utils.hpp"
 
-#include "restcl_definitions.hpp"
 #include "http_frame.hpp"
